pyproject.toml
[project]
name = "tetrika-test-task"
version = "0.1.0"
requires-python = ">=3.13.3, <4.0"
dependencies = [
    "aiohttp>=3.12.2",
    "python-dotenv>=1.1.0",
    "wikipedia-api>=0.8.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[dependency-groups]
dev = [
    "autoflake>=2.3.1",
    "black>=25.1.0",
    "isort>=6.0.1",
    "mypy>=1.15.0",
    "pytest>=8.3.5",
]

# --------------------------------------------------------------------------------------


# Code formatting

[tool.autoflake]
in_place = true
recursive = true
remove_all_unused_imports = true

[tool.black]
line-length = 88

[tool.isort]
force_grid_wrap = 2
include_trailing_comma = true
line_length = 88
lines_after_imports = 2
multi_line_output = 3

# --------------------------------------------------------------------------------------


# Linting

[tool.mypy]
exclude = "test"
explicit_package_bases = true
ignore_missing_imports = true
mypy_path = "src"
strict = true
